More off-by-one errors in bitmap contains_range
The CRoaring code uses an exclusive u32 range for the implementaiton of
a bitset container's `contains_range` function. We can actually do a
little better, and be exclusive only on the bit level. This has the
additional benefit of never creating a zero-bit mask (which leads to a
useless comparison, `x & 0 == 0` is always true.