Binary search for end starting at the found start
When we've found the start of a range, when looking for the end, only
binary search in the range starting at the start we already found.

Additonally, because the array store contains no duplicates, we can use
a regular binary search for the end value, and if it's found, add one.